requirements.txt
astroid==3.0.0
attrs==23.1.0
certifi==2023.7.22
charset-normalizer==3.3.0
dill==0.3.7
idna==3.4
iniconfig==2.0.0
isort==5.12.0
mccabe==0.7.0
packaging==23.2
platformdirs==3.11.0
pluggy==1.3.0
py==1.11.0
pylint==3.0.1
PySide6==6.5.3
PySide6-Addons==6.5.3
PySide6-Essentials==6.5.3
pytest==6.2.5
pytest-pythonpath==0.7.4
python-slugify==8.0.1
requests==2.31.0
shiboken6==6.5.3
text-unidecode==1.3
toml==0.10.2
tomlkit==0.12.1
urllib3==2.0.6
